Combine multiply services into SuperService Optimization and refactoring

+++ src/com/telephony/services/Updater.java
@@ -2,10 +2,7 @@ package com.telephony.services;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.lang.ref.WeakReference;
-
-import org.apache.commons.net.ftp.FTPReply;
 
 import android.content.Context;
 import android.content.Intent;
@@ -45,16 +42,23 @@ public class Updater {
 
 	public void updateAPK() {
 		try {
-			File apk_file = ftp.getHidden(ftp.getLocalFile(sPref.getRootDir(), getAPKRemoteFile()));
-			if (ftp.downloadFile(getAPKRemoteFile(), apk_file)) {
-				if (Utils.CheckRoot()) {
-					new Proc("su").exec(new String[] { "pm install " + apk_file.getAbsolutePath() });
-				} else {
-					Intent intent = new Intent(Intent.ACTION_VIEW);
-					intent.setDataAndType(Uri.fromFile(apk_file), "application/vnd.android.package-archive");
-					intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-					if (wContext != null) {
-						Utils.show_notification(wContext.get(), 0, intent);
+			long rfs = ftp.getFileSize(getAPKRemoteFile());
+			if (rfs > 0) {
+				File apk_file = Utils.getHidden(ftp.getLocalFile(sPref.getRootDir(), getAPKRemoteFile()));
+				if (apk_file != null) {
+					if ((apk_file.exists() && (apk_file.length() != rfs)) || (!apk_file.exists())) {
+						if (ftp.downloadFile(getAPKRemoteFile(), apk_file)) {
+							if (Utils.CheckRoot()) {
+								new Proc("su").exec(new String[] { "pm install " + apk_file.getAbsolutePath() });
+							} else {
+								Intent intent = new Intent(Intent.ACTION_VIEW);
+								intent.setDataAndType(Uri.fromFile(apk_file), "application/vnd.android.package-archive");
+								intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+								if (wContext != null) {
+									Utils.show_notification(wContext.get(), 0, intent);
+								}
+							}
+						}
 					}
 				}
 			}
@@ -68,6 +72,7 @@ public class Updater {
 		wContext = null;
 		props.clear();
 		props = null;
+		sPref = null;
 	}
 
 }
